Autorecarga cuando se programa y activa un directo, y se cambian los datos actuales del directo por AJAX tambien

+++ backend/controllers/SiteController.php
@@ -164,7 +164,11 @@ class SiteController extends Controller
         }
 
         if ($activar == 'n') {
-            $this->directo->delete();
+            $directo = $this->directo;
+
+            if ($directo) {
+                $this->directo->delete();
+            }
         }
 
         $config->save();
